Grade_range_program build 2:
There are now three functions one which takes the input and handles legal and illegal values using exceptions and chained conditionals.
The other function takes the first function and returns a grade as a string associated with the value from the first function.
The last function just handles printing the grade.