Backwards_string_program build 2:
The program now handles the exceptions of integers through a error handling function that checks if there are integers
in the input then raises a ValueError which prints an error message. This will also return a different value that
is checked for in the main function and breaks the loop if the value returned matches the if statement.

The program continues as normal when entering a legal variable. Still is the program doesn't print letters on the
same line in the terminal.